Some major changes to OLED
Had to do some major changes to the OLED function because the functions
were not behaving properly. Made a prototype for the Menu system and
organized the struct for it and some functions. We've got to make a
Three structure menu that we can iterate through and print it out on the
screen.